Refactor wiki module
