+++ app/__init__.py
@@ -6,6 +6,7 @@ from . import session
 from . import bbcode
 from . import common
 from . import routes
+from . import wiki
 
 from .app import flask, get_handle
